// src/mission_members_payments/dto/mission_members_payments.dto.ts
export class creatememberDto {
    Mission_ID: number;
        Staff_Id: number;
        Done_On: Date;
        Beneficially_Bank: string;
        Account: string;
        Amount: number;
        Done_By: number;
        autholized_by: number;
        Mission_Periode_Days: number;
        Percentage_of_Payment: number;
}

export class updatememberDto {
    Mission_ID: number;
        Staff_Id: number;
        Done_On: Date;
        Beneficially_Bank: string;
        Account: string;
        Amount: number;
        Done_By: number;
        autholized_by: number;
        Mission_Periode_Days: number;
        Percentage_of_Payment: number;
}